requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
drf-spectacular==0.27.2
django-filter==23.5
django-extensions==3.2.3
pydotplus==2.0.2
